+++ openGlTextures/buffers/ElementBuffer.h
@@ -4,8 +4,8 @@ class ElementBuffer
 {
 public:
 	unsigned int bufferId;
-	ElementBuffer(const float* indices, unsigned int size);
-	 const void bind();
+	ElementBuffer(const unsigned int* indices, size_t size);
+	const void bind();
 	 const void unbind();
 };
